Refactor RAG test fixtures into reusable plugin
